+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- In Cypress, Added optional Order Suffix to reference number testcase
+- [ENGINEERS-1138] - In Cypress, Added optional Order Suffix to reference number testcase
 
 ## [1.13.9] - 2023-02-03